refactor DU monitors page
